feat: create crud for users and tests

+++ model/goal.go
@@ -6,12 +6,12 @@ import (
 )
 
 type Goal struct {
-	ID       int         `json:"id"`
+	ID       int64         `json:"id"`
 	Name     string      `json:"name"`
 	Desc     string      `json:"description,omitempty"`
 	Price    utils.Money `json:"price"`
 	Pros     *string     `json:"pros,omitempty"`
 	Cons     *string     `json:"cons,omitempty"`
-	UserID   int         `json:"user_id"`
+	UserID   int64         `json:"user_id"`
 	Deadline time.Time   `json:"deadline"`
 }
