+++ model/transaction.go
@@ -6,10 +6,10 @@ import (
 )
 
 type Transaction struct {
-	ID        int         `json:"id"`
+	ID        int64         `json:"id"`
 	Desc      *string     `json:"description"`
 	Amount    utils.Money `json:"amount"`
 	IsDebt    bool        `json:"is_debt"`
 	CreatedAt time.Time   `json:"created_at"`
-	UserID    int         `json:"user_id"`
+	UserID    int64         `json:"user_id"`
 }
